refactor: move global types to types/index.ts
